\name{animals}
\alias{animals}
\non_function{}
\title{Survival Times of Animals}
\description{
The \code{animals} data frame records the survival times of 48 animals given different poisons and treatments.
}
\format{
This data frame has 48 observations on 3 variables:
 \tabular{lll}{
  Poison \tab 
  factor \tab levels
  \code{I} 
  \code{II} 
  \code{III} 
  \cr
  Treatment \tab
  factor \tab levels
  \code{A} 
  \code{B} 
  \code{C} 
  \code{D} 
  \cr
  Survival \tab numeric \tab animal's survival time 
  measured in 10 hour units
 }
}
\source{
Handbook of Small Datasets; originally in Box and Cox (1964), JRSS A
volume 143, pages 383-430.
}
\keyword{datasets}
