\name{byssinosis}
\alias{byssinosis}
\non_function{}
\title{Byssinosis Incidence in US Cotton Industry Workers}
\description{
The data come from a survey of workers in the US cotton industry and
record whether they were suffering from the lung disease byssinosis as
well as the values of five explanatory variables: the race, sex and
smoking status of the worker, the length of employment, and the
dustiness of the workplace.
}
\format{
  A data frame with 65 rows and 8 columns:
  \tabular{rlll}{
    [,1] \tab Yes \tab numeric \tab Number of workers in this category with byssinosis \cr
    [,2] \tab No \tab numeric \tab Number of workers in this category without byssinosis \cr
    [,3] \tab Dust \tab factor \tab 1 (high), 2 (medium) or 3 (low) \cr
    [,4] \tab Race \tab factor \tab 1 (white) or 2 (other) \cr
    [,5] \tab Sex \tab  factor \tab 1 (male) or 2 (female) \cr
    [,6] \tab Smoking \tab factor \tab 1 (smoker) or 2 (non-smoker) \cr
    [,7] \tab Emp.length \tab factor \tab 1 (< 10 years), 2 (10-20
    years) or 3 (> 20 years) \cr
    [,8] \tab n \tab numeric \tab  number of workers in this category, i.e. Yes+No
  }
}
\source{
Handbook of small datasets (originally in Higgins and Koch (1977),
International Statistical Review, volume 45, pages 51-62.
}
\keyword{datasets}
