\name{chlorph}
\alias{chlorph}
\title{Amounts of Chlorpheniramine Maleate in Tablets}
\description{
The data come from a semiautomated process for measuring the actual amount of chlorpheniramine maleate in tablets nominally containing 4mg. The tablets used for the study were made by two different manufacturers.  For each manufacturer, a composite was produced by grinding together a number of tablets.  Each composite was split into seven pieces each of the same weight as a tablet and the pieces were sent to seven different laboratories. Each laboratory made 10 separate measurements on each composite.
}
\usage{data(chlorph)}
\format{
  A data frame with 140 rows and 3 columns:
  \tabular{rlll}{
    [,1] \tab chlorpheniramine \tab numeric \tab Amount measured \cr
    [,2] \tab manufacturer \tab factor \tab Tablet manufacturer (A or B)\cr
    [,3] \tab laboratory \tab factor \tab Laboratory where measured (1 to 7)
  }
}
\source{PSC can't remember}
 
\keyword{datasets}
