\name{cork}
\alias{cork}
\non_function{}
\title{Cork Deposits}
\description{
The \code{cork} data record
the weight of cork deposits for 28 trees, in each of 4 directions.
}
\format{
A data frame with 28 observations on 4 variables, each of which 
measures the weight (centigrams) of cork.\cr
  \tabular{rlll}{
  [,1] \tab N \tab numeric \tab Cork on the North side of the tree \cr
  [,2] \tab E \tab numeric \tab Cork on the East side of the tree \cr
  [,3] \tab S \tab numeric \tab Cork on the South side of the tree \cr
  [,4] \tab W \tab numeric \tab Cork on the West side of the tree  
  }
}
\source{
Handbook of Small Datasets; originally from Rao (1948), Biometrika,
volume 35, pages 58--79.
}
\keyword{datasets}
