\name{crime}
\alias{crime}
\non_function{}
\title{US Crime Statistics}
\description{
Numbers are given of offences known to police per 100,000 residents of
fifty states of the USA plus the District of Columbia for the year
1986.
}
\format{
  A data frame with 7 columns:
  \tabular{rl}{
    [,1] \tab Murder \cr
    [,2] \tab Rape \cr
    [,3] \tab Robbery \cr
    [,4] \tab Agg.Assault \cr
    [,5] \tab Burglary \cr
    [,6] \tab Larceny \cr
    [,7] \tab Vehicle.Theft
    }
}
\source{
Handbook of Small Datasets (originally from the Statistical Abstract of the US (1988), volume 150, Table 265.
}
\keyword{datasets}
