\name{decay}
\alias{decay}
\title{Radioactive Decay}
\description{
The data come from an experiment performed by Meyer and Von Schweidler in 1905 concerning radioactive decay. The data represent the time in days, and the degree of radioactivity measured at that time for a particular material. 
}
\usage{data(decay)}
\format{
  A data frame with 14 rows and two columns: 
  \tabular{rlll}{
    [,1] \tab Days \tab numeric \tab Days since beginning of experiment \cr
    [,2] \tab Activity \tab numeric \tab Level of radioactivity
  } 
}
\source{JW Tukey, "Exploratory Data Analysis", chapter 6.}

\keyword{datasets}
