\name{engine}
\alias{engine}
\title{Pollutants from Engines}
\description{
  The data record the amounts of three pollutants emitted by various light-duty engines.  
}
\usage{data(engine)}
\format{
  A data frame with 46 rows and 3 columns.
  \tabular{rlll}{
    [,1] \tab CO \tab numeric \tab Carbon monoxide emitted per mile (g) \cr
    [,2] \tab HC \tab numeric \tab Hydrocarbons emitted per mile (g) \cr
    [,3] \tab NOX\tab numeric \tab Nitrogen oxide emitted per mile (g) 
  }
}
\source{
  Table 2.9 of Moore and McCabe, "Introduction to the Practice of  Statistics".
}
\keyword{datasets}
