\name{hips}
\alias{hips}
\title{Wear in Replacement Hip Joints}
\description{
At some time during their lives, many people will have one or both of their hip joints replaced by an artificial joint. Such replacements are increasingly successful and long-lasting. However, there is still much interest in understanding the processes which result in failure of joints. One such process is the wear that takes place to the surface of the parts of the joint. The data record, for a number of joints recovered from patients, the length of time that each joint was implanted in the patient's body and the depth of the wear to the socket of the joint. It is desired to understand the relationship between implantation time and wear depth.
}
\usage{data(hips)}
\format{
  A data frame with 124 rows and two columns: 
  \tabular{rlll}{
    [,1] \tab Time \tab numeric \tab How long the joint was implanted (years) \cr
    [,2] \tab Depth \tab numeric \tab Depthe of wear at removal from patient
  }
}
\source{Richard Hall, University of Leeds.}

\keyword{datasets}
