\name{hospital}
\alias{hospital}
\title{Patient discharges from short-stay hospitals}
\description{
The data comprise the numbers of discharges for 393 short-stay
hospitals during January 1968.
}
\format{This data frame contains the following columns:
 \describe{
    \item{beds}{number of beds}
    \item{discharges}{number of discharges during January 1968}
    \item{region}{unspecified geographical region}
 }
}
\source{J. A. Rice (1985). \emph{Mathematical Statistics and Data Analysis
(2nd Edition)}, p186. Belmont, California: Duxbury Press.}
\examples{
data(hospital)
}
\keyword{datasets}
