\name{icecream}
\alias{icecream}
\title{Ice Cream Sales}
\description{
The data were collected in an attempt to understand what affects the consumption of ice cream.  Ice cream consumption was measured over 30 four-week periods from March 18th 1951 to July 11th 1953. It was thought that variables influencing consumption might include the price of ice cream, weekly family income and the temperature.
}
\usage{data(icecream)}
\format{
  A data frame with 30 rows and 4 columns: 
  \tabular{rlll}{
    [,1]\tab Consumption \tab numeric\tab amount consumed (pints per capita)\cr
    [,2]\tab Price \tab numeric\tab price of ice cream (dollars per pint)\cr
    [,3]\tab Income \tab numeric\tab weekly family income (dollars)\cr
    [,4]\tab Temperature \tab numeric\tab mean temperature (Fahrenheit)
  }
}
\source{Some introductory stats book? PSC can't remember}
\keyword{datasets}
