\name{insect}
\alias{insect}
\title{Measurements of Insect Joints}
\description{
The data are measurements of the size of three joints for 10 insects
from each of three different species of Chaetocnema.
}
\usage{data(insect)}
\format{
  A data frame with 30 rows and 4 columns: 
  \tabular{rlll}{
    [,1]\tab X1 \tab numeric \tab Measurement of joint 1 \cr
    [,2]\tab X2 \tab numeric \tab Measurement of joint 2 \cr
    [,3]\tab X3 \tab numeric \tab Measurement of joint 3 \cr
    [,4]\tab Species \tab factor \tab Species of insect (A,B, or C)
  }
}
\source{Handbook of Small Datasets.}
\keyword{datasets}
