\name{leghorn}
\alias{leghorn}
\non_function{}
\title{White Leghorn Fowl}
\description{
The data concern 6 bone measurements for 275 white leghorn fowl. The
measurements are of skull length, skull breadth, and lengths of the
humerus, ulna, femur and tibia.
}
\format{Correlation matrix between the 6 bone measurements}
\source{Allan Seheult}
\keyword{datasets}
