\name{leukaemia}
\alias{leukaemia}
\non_function{}
\title{Remission times for leukaemia patients}
\description{
The data are the remission times for 21 leukaemia patients who were
the control group in a clinical trial for a treatment for the condition.
}
\format{
A vector of 21 times
}
\source{
I forget
}
\keyword{datasets}
