\name{lightspeed}
\alias{lightspeed}
\title{Michelson's speed of light determinations}
\description{
The data comprise 100 observations representing Michelson's
determinations of the velocity of light made from 2 June to 2 July
1879.
}
\format{A vector of measurements representing the velocity of
light minus 299,000.}

\source{J. A. Rice (1995). \emph{Mathematical Statistics and Data Analysis
(2nd Edition)}, p324. Belmont, California: Duxbury Press.}
\examples{
data(lightspeed)
}
\keyword{datasets}

