\name{lunatics}
\alias{lunatics}
\title{Lunatics in Massachusetts Asylums, 1854}
\description{
These data are from an 1854 survey conducted by the Massachusetts Commission on Lunacy under the leadership of Edward Jarvis. Dr. Jarvis was President of the American Statistical Association from 1852 to 1882. They were influential in changing policy in relation to asylums. Five variables are recorded for each of 14 counties in the state of Massachusetts.
}
\usage{data(lunatics)}
\format{
  A data frame with 14 rows and 5 columns:
  \tabular{lll}{
    Number \tab numeric \tab Number of lunatics in the county \cr
    Distance \tab numeric \tab Distance to nearest mental health center \cr
    Population \tab numeric \tab Population of county (thousands) \cr
    Density \tab numeric \tab Density of population (per square mile) \cr
    Home \tab numeric \tab Percentage of lunatics cared for at home
  }
}
\source{Handbook of Small Datasets?}
\keyword{datasets}
