\name{mice}
\alias{mice}
\title{Agressiveness of Mice}
\description{
Three species of mice were tested for ``aggressiveness''. The species were A/J, C57 and F2 (a cross of the first two species).  Each mouse was placed in a 1-meter-square box, which was marked off into 49 equal squares.  The mouse was let go on the center square, and the number of squares traversed in a five-minute period was counted.
}
\usage{data(mice)}
\format{
  A data frame with 134 rows and 2 columns: 
  \tabular{lll}{
     species \tab factor \tab Species of mouse \cr
     squares \tab numeric \tab Number of squares traversed
  }
}
\source{J.A. Rice, Mathematical Statistics and Data Analysis (1988).}
\keyword{datasets}
