\name{missile}
\alias{missile}
\non_function{}
\title{Lifetimes of Missile Batteries}
\description{
The data record result from a factorial experiment with four replicates 
conducted to investigate the effects of two factors, ambient temperature and
battery plate material, on the effective lifetime of batteries used
for shoulder-fired ground-to-air missiles.
}
\usage{data(missile)}
\format{A data frame with 3 variables:
\tabular{lll}{
Temperature \tab factor \tab Low, Medium or High \cr
Material \tab factor \tab 1, 2 or 3 \cr
Battery.life \tab numeric \tab Battery life (hours) \cr
}}
\source{
	Montgomery and Runger, p.710, problem 12-1. Originally from:
	DC Montgomery, Design and Analysis of Experiments, 3rd edition, 
	Wiley, 1991.
}
\keyword{datasets}
