\name{moths}
\alias{moths}
\non_function{}
\title{Numbers of Lepidoptera in Rothamstead light trap}
\description{
The data are the number of Lepidoptera of various species trapped in a
light trap at Rothamstead from 1933-1936
}
\format{
  A data frame with 2 columns:
  \tabular{rlll}{
    [,1] \tab Frequency \tab numeric \tab number of species for which a particular number of specimens were trapped \cr
    [,2] \tab Size \tab numeric \tab number of specimens trapped of each of those species
    }
}
\source{
Fisher, Corbett and Williams, J. Animal Ecology, 12, pages 42 to 58.
}
\keyword{datasets}
