\name{octopod}
\alias{octopod}
\title{Dorsal lengths of octopods}
\description{
The data are the dorsal lengths of taxonomically distinct octopods.
}
\usage{data(octopod)}
\format{
  A data frame with 94 rows and 1 column:
  \tabular{lll}{
    dorsal \tab numeric \tab Dorsal length (millimetres) 
  }
}
\source{David Wooff}
\keyword{datasets}
