\name{plasma}
\alias{plasma}
\title{Cholesterol levels}
\description{
A group of 24 patients with hyperlipoproteinaemia, a metabolic disorder
characterised by high levels of lipoproteins in the blood that can be
associated with coronary heart disease are observed to have the
following plasma levels.
}
\format{This data frame contains the following columns:
 \describe{
    \item{age}{age of patient in years}
    \item{cholesterol}{plasma level in total cholesterol (mg/ml)}
 }
}
\source{W. K. Krzanowski (1998). \emph{An Introduction to Statistical
Modelling}, p51. London: Arnold.}
\examples{
data(plasma)
}
\keyword{datasets}
