\name{protein}
\alias{protein}
\non_function{}
\title{European Protein Consumption}
\description{
The data record protein consumption for nine food groups in
twenty-five European countries.  The data relate to a period sometime
in the early 1970s.
}
\format{
  A data frame with 9 columns:
  \tabular{rl}{
    [,1] \tab Red.meat\cr
    [,2] \tab White.meat\cr
    [,3] \tab Eggs\cr
    [,4] \tab Milk\cr
    [,5] \tab Fish\cr
    [,6] \tab Cereals\cr
    [,7] \tab Starchy.foods\cr
    [,8] \tab Pulses.nuts.seeds\cr
    [,9] \tab Fruit.veg\cr
    }
}
\source{
Handbook of Small Datasets, 360. Originally from a publication in
German by Weber (1973). Original English language publication by
Gabriel (1981) in Interpreting Mulvariate Data (ed. V. Barnett), pages
147-173.
}
\keyword{datasets}
