\name{seaheight}
\alias{seaheight}
\non_function{}
\title{Annual Maximum Sea Height at Immingham}
\description{
The data are the annual maximum sea level at Immingham from 1920 to
1988. The units and detail of what is measured have been mislaid.
}
\format{A time-series of annual measurements}
\source{Mark Dixon (formerly at  University of Newcastle upon Tyne)}
\keyword{datasets}

