\name{senility}
\alias{senility}
\title{Wechsler scores and senility in elderly men}
\description{
Forty-nine elderly men participating in an interdisciplinary study of aging were classified into the diagnostic categories ``senile factor present'' (12 subjects) and ``no senile factor'' (37 subjects) on the basis of an intensive psychiatric examination. The Wechsler Adult Intelligence Scale (WAIS) had been administered to all subjects by an independent investigator, and certain sub-tests (information, similarities, arithmetic, and picture completion) showed large differences between the two groups.
}
\usage{data(senility)}
\format{
  A data frame with 49 rows and five columns: 
  \tabular{lll}{
    Psychiatric \tab factor \tab Result of psychiatric examination \cr
    WSimilarities \tab numeric \tab Score on Wechsler similarities test \cr
    WArithmetic \tab numeric \tab Score on Wechsler arithmetic test \cr
    WPicture \tab numeric \tab Score on Wechsler picture completion test \cr
    WInformation \tab numeric \tab Score on Wechsler information test
  }
}
\source{D.F. Morrison, Multivariate Statistical Methods.}
\keyword{datasets}
