\name{shuttle}
\alias{shuttle}
\non_function{}
\title{O-Ring Failures on the Space Shuttle}
\description{
The data record the number of O-ring failures and the temperature at
launch for every space shuttle flight prior to the Challenger
disaster.
}
\format{
  A data frame with 2 columns:
  \tabular{rlll}{
    [,1] \tab Temperature \tab numeric \tab launch temperature in Fahrenheit \cr
    [,2] \tab Failures \tab numeric \tab number of O-rings which failed
    (out of possible 6)
  }
}
\source{Not known.
}
\keyword{datasets}
