\name{soapsuds}
\alias{soapsuds}
\title{Height of soap suds}
\description{
The data are from an experiment in which the weight of soap was
varied, and the height of suds was measured subject to a given amount
of agitation.
}
\format{This data frame contains the following columns:
 \describe{
    \item{soapweight}{weight of soap in gm}
    \item{sudheight}{height of suds in cm}
}
}
\source{W. K. Krzanowski (1998). \emph{An Introduction to Statistical
Modelling}, p105. London: Arnold.}
\examples{
data(soapsuds)
}
\keyword{datasets}
