\name{statcoll}
\alias{statcoll}
\title{Statistics Collection Marks 1996}
\description{
The data are the marks obtained by students taking S Statistics in 1995-6 on each of the four questions on the collection examination paper sat in the first week of term after Christmas.
}
\usage{data(statcoll)}
\format{
  A data frame 56 rows and 4 columns: 	
  \tabular{lll}{
    q1 \tab numeric \tab Mark for question 1 \cr
    q2 \tab numeric \tab Mark for question 2 \cr
    q3 \tab numeric \tab Mark for question 3 \cr
    q4 \tab numeric \tab Mark for question 4
  }
}
\source{Peter Craig}
\keyword{datasets}
