\name{stud0001}
\alias{stud0001}
\title{Student ID numbers of DAMS and Statistics students 2000-01}
\description{
  The data record Durham student identification numbers of DAMS and Statistics students for the academic year 2000-2001 and which module each student was taking.
}
\usage{data(stud0001)}
\format{
	A data frame with 75 rows and 2 columns.
	\tabular{rlll}{
	[,1] \tab id \tab numeric \tab Student's ID number\cr
	[,2] \tab stats \tab boolean \tab  TRUE if taking Statistics
	}
}
\source{Peter Craig <P.S.Craig@durham.ac.uk>}
\keyword{datasets}

