\name{stud9900}
\alias{stud9900}
\title{Student ID numbers of DAMS and Statistics students 1999-00}
\description{
  The data record Durham student identification numbers of DAMS and Statistics students for the academic year 1999-2000 and which module each student was taking.
}
\usage{data(stud9900)}
\format{
	A data frame with 86 rows and 2 columns.
	\tabular{rlll}{
	[,1] \tab id \tab numeric \tab Student's ID number\cr
	[,2] \tab stats \tab boolean \tab  TRUE if taking Statistics
	}
}
\source{Peter Craig <P.S.Craig@durham.ac.uk>}
\keyword{datasets}

