\name{student}
\alias{student}
\title{Students in S Statistics 1993-4}
\description{
  Personal characteristics are recorded for students taking S Statistics in 1993-4. Fitness level was recorded on an informal scale, the details of which now elude me. The important information is that  higher numbers are supposed to represent greater levels of fitness.
}
\usage{data(student)}
\format{
  A data frame with 55 rows and 8 columns.
  \tabular{rlll}{
    [,1] \tab Gender \tab numeric \tab gender coded as 1 or 0 \cr
    [,2] \tab Height \tab numeric \tab height in inches"      \cr
    [,3:4] \tab Stone, Lb \tab numeric \tab weight in stone and pounds \cr
    [,5] \tab Shoe   \tab numeric \tab shoe size (British)    \cr
    [,6:7] \tab Years, Months \tab numeric \tab age in years and months\cr
    [,8] \tab Fitness\tab numeric \tab self-assessed fitness level
  }
}
\source{Peter Craig <P.S.Craig@durham.ac.uk>}
\keyword{datasets}
