\name{ucas9899}

\alias{ucas9899}

\title{UCAS numbers of DAMS and Statistics students 1998-1999}

\description{
The data record UCAS numbers of DAMS and Statistics students for
the academic year 1998-99 and which module each student was
taking.
}

\usage{data(ucas9899)}

\format{
A data frame with 102 rows and 2 columns.
\tabular{rlll}{
	[,1] \tab ucasno \tab numeric \tab Student's UCAS number\cr
	[,2] \tab stats \tab boolean \tab  \code{TRUE} if taking Statistics
}
}
\source{Peter Craig <P.S.Craig@durham.ac.uk>}
\keyword{datasets}
