\name{voles}
\alias{voles}
\non_function{}
\title{Water Vole Skulls}
\description{
The data is a dissimilarity matrix for 14 European populations of
water vole. 
}
\format{A dissimilarity matrix}
\details{ The original data (not given here) record the presence or
absence of 13 characteristics in about 300 water vole skulls divided
into samples from 14 populations from Britain and the rest of
Europe. The dissimilarity matrix given here was computed from the
original data.}
\source{Handbook of Small Datasets, 34. Originally from Corbet et al (1970),
Journal of Zoology, volume 161, pages 301-316.}
\keyword{datasets}

