\name{whales}
\alias{whales}
\non_function{}
\title{Whales' 1km Swimming Times}
\description{
The data are estimates of the time that each of 210 whales would take
to swim 1km. Pairs of sightings and corresponding locations that could
be reliably attributed to the same whale were collected, thus
providing an estimate of velocity for each whale. The velocities were
converted into times taken to swim 1km.
}
\format{A vector of 210 times.}
\source{Rice, Mathematical Statistics and Data Analysis 2nd ed,
p.292. Originally from Raftery and Zeh (1993) Estimation of bowhead
whales, Balaena mysticetus, population size. In Gatsonis et al (eds),
Case Studies in Bayesian Statistics, Springer LNS 83.}
\keyword{datasets}
