\name{zeamays}
\alias{zeamays}
\title{Darwin's Zea Mays plant data}
\description{
This dataset was produced by Charles Darwin.  Darwin was exploring the effects
of in-breeding as opposed to cross-breeding on the "strength" of species.  One
of his experiments was performed on Zea Mays plants.  Several pairs of seeds
were taken from Zea Mays plants grown under controlled conditions.  The two
seeds in each pair were taken from the same plant; however, one of the pair was
taken from a self-fertilised flower (one fertilised by the same plant) and the
other from a cross-fertilised flower (one fertilised by a different plant).
The two seeds were subsequently grown together under identical conditions and
their heights (in inches) were measured.
}
\usage{data(learndis)}
\format{
  A data frame with 15 rows and 3 columns:
  \tabular{ll}{
    Batch \tab indicates which group of Darwin's plants the pair came from \cr
    Cross \tab heights (inches) of the cross-fertilised plants \cr
    Self \tab heights (inches) of the self-fertilised plants
  }
}
\source{Darwin, C. (1876). The Effect of Cross- and Self-fertilization in the Vegetable Kingdom, 2nd Ed. London: John Murray.}
\keyword{datasets}
